RFC 2.1
With "RFC", you can easily compare two strings and check if they are equal or not.

With "RFC 2", you don't even have to use the Enter key or push the button. The strings get compared as you enter them. Version 2.0 had some bugs, so be sure to use version 2.1 instead.

Both versions check if there are spaces in front or behind the strings that could have been copied unintendedly.